feat: Replace the existing Resources search with Azure AI Search  (#356)
* Fix the config key

* Initial SearchController

* Add new package

* Update CSS for search

* Search View (wip)

* Update the clear filters function

* Update filter categories

* Order by js to submit form

* Css changes

* CSS update

* Search refactor

* Fix the application configuration tests

* Add SearchResultsVMFactoryTests

* Fix tests

* Create SearchResourcesControllerTests.cs

* Remove old resources search and substitute new one in

* Add resource search facet numbers to filter control

* Make resource search filters inclusive OR instead of exclusive AND

* Change the gds filter checkbox size to match the design

* Update CSS

* Display resource properties in search results and resource page

* Small refactor

* Update SearchResources.cshtml

* Remove unused code

* Change the query string parameter names and logically OR tags instead of AND

* Tidy up old Resource search code

* Sonar suggestions

* Remove special characters from the search term

* Update SearchServiceTests.cs

* Set a max query size

* Delete search-sort.js

* Handle null query

* Update SearchResultsVMFactory.cs

* Update README.md

* Update docker-compose.yml

* Update Resource.cshtml

+++ Childrens-Social-Care-CPD/TagHelpers/GovUkPaginationTagHelper.cs
@@ -13,10 +13,10 @@ public class GovUkPaginationTagHelper : TagHelper
     public string UrlFormatString { get; set; }
 
     [HtmlAttributeName("page-count")]
-    public int PageCount { get; set; }
+    public long PageCount { get; set; }
 
     [HtmlAttributeName("current-page")]
-    public int CurrentPage { get; set; }
+    public long CurrentPage { get; set; }
 
     public override void Process(TagHelperContext context, TagHelperOutput output)
     {
@@ -44,7 +44,7 @@ public class GovUkPaginationTagHelper : TagHelper
         }
     }
 
-    private static IHtmlContent GeneratesPagesPages(int pageCount, int currentPage, string formatString)
+    private static IHtmlContent GeneratesPagesPages(long pageCount, long currentPage, string formatString)
     {
         var ul = new TagBuilder("ul");
         ul.AddCssClass("govuk-pagination__list");
